move split_data to utils
