+++ grants_tagger_light/preprocessing/preprocess_mesh.py
@@ -14,8 +14,7 @@ import yaml
 from tqdm import tqdm
 
 from grants_tagger_light.label_binarizer import create_label_binarizer
-from grants_tagger_light.split_data import split_data
-from grants_tagger_light.utils import verify_if_paths_exist, write_jsonl
+from grants_tagger_light.utils import verify_if_paths_exist, write_jsonl, split_data
 
 
 def yield_raw_data(input_path):
