+++ grants_tagger_light/utils/__init__.py
@@ -1,3 +1,10 @@
 from .utils import load_data, load_train_test_data, verify_if_paths_exist, write_jsonl
+from .split_data import split_data
 
-__all__ = ["load_train_test_data", "load_data", "verify_if_paths_exist", "write_jsonl"]
+__all__ = [
+    "load_train_test_data",
+    "load_data",
+    "verify_if_paths_exist",
+    "write_jsonl",
+    "split_data",
+]
